Upgrade the concept of 'graphics-logic API' to a general purpose event-driven mecanism
The change came because, for instance, a logical event
can be connected to logical callbacks besides the graphical
ones. Of course, the older events are a subset of the newer.
This way is more flexible but, maybe, harder to debug and
trace. Now, there are two static classes, Logic and Graphics,
the former contains the logical events while the latter
contains the graphical events. Both classes are defined
inside the namespace Events.